Move render doctest to separate test file
